Add collaborators to environments
I forgor to commit

+++ src/lib/server/auth.ts
@@ -3,6 +3,12 @@ import { TokenModel } from "./models";
 import { dev } from "$app/environment";
 import { randomBytes } from 'crypto';
 
+export async function removeExpiredTokens(){
+    await TokenModel.deleteMany({ refreshTokenExpiration: {
+        $lt: Date.now()
+    }});
+}
+
 export async function getUserFromCookie(cookies: Cookies) {
     const authCookie = cookies.get("auth-token");
     const refreshCookie = cookies.get("refresh-token");
@@ -10,12 +16,18 @@ export async function getUserFromCookie(cookies: Cookies) {
         return undefined;
     }
 
+    await removeExpiredTokens();
+
     let token = await TokenModel.findOne({ authToken: authCookie }).populate('user');
 
     // if token is expired
     if (token == undefined || Date.now() > token.authTokenExpiration.getTime()) {
         token = await TokenModel.findOne({ refreshToken: refreshCookie }).populate('user');
         if (token == undefined) return undefined;
+
+        if(Date.now() > token.refreshTokenExpiration) {
+            return undefined;
+        }
 
         // generate new token and delete old one
         await TokenModel.deleteOne({ _id: token._id });
@@ -66,7 +78,7 @@ export async function generateToken(user: any) {
         authToken: generateRandomToken(),
         refreshToken: generateRandomToken(),
         authTokenExpiration: Date.now() + (1000 * 60 * 60),
-        expiresAt: Date.now() + (1000 * 60 * 60 * 24 * 30)
+        refreshTokenExpiration: Date.now() + (1000 * 60 * 60 * 24 * 30)
     });
 
     await token.populate('user');
